feat(UserValues) : move Values for user from Lotto config
