--- src/main/java/lotto/ErrorMessages.java
@@ -1,4 +1,0 @@
-package lotto;
-
-public enum ErrorMessages {
-}
