+++ src/main/java/machines/AutoLottoMachine.java
@@ -2,7 +2,6 @@ package machines;
 
 import static config.SettingValues.*;
 
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
