+++ src/test/java/ApplicationTest.java
@@ -1,6 +1,6 @@
-import camp.nextstep.edu.missionutils.test.NsTest;
-import lotto.Application;
+package lotto;
 
+import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
